+++ stroom-core/src/main/java/stroom/core/entity/cluster/ClearableResourceImpl.java
@@ -16,11 +16,9 @@
 
 package stroom.core.entity.cluster;
 
-import com.codahale.metrics.health.HealthCheck.Result;
 import stroom.node.api.NodeCallUtil;
 import stroom.node.api.NodeInfo;
 import stroom.node.api.NodeService;
-import stroom.util.HasHealthCheck;
 import stroom.util.jersey.WebTargetFactory;
 import stroom.util.shared.ResourcePaths;
 
@@ -30,7 +28,7 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.util.Objects;
 
-class ClearableResourceImpl implements ClearableResource, HasHealthCheck {
+class ClearableResourceImpl implements ClearableResource {
     private final NodeService nodeService;
     private final NodeInfo nodeInfo;
     private final WebTargetFactory webTargetFactory;
@@ -74,10 +72,5 @@ class ClearableResourceImpl implements ClearableResource, HasHealthCheck {
                 throw NodeCallUtil.handleExceptionsOnNodeCall(nodeName, url, e);
             }
         }
-    }
-
-    @Override
-    public Result getHealth() {
-        return Result.healthy();
     }
 }